pocket-integration/requirements.txt
nats-py==2.1.4
pocket-api==0.1.5
